basic_adk_agent/pyproject.toml
[project]
name = "basic"
version = "0.1"
description = "Basic agent"
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
  "google-adk>=1.8.0",
]